bcstatetransfer_tests: Add new infrastructure class and fake replicas
1) DataGenerator: Generates blocks, reserved pages, blocks, digests, and
whatever other data is needed for the test to run. Part of this class is
composed of code being refactored out from the fake source code, and
part of it is new written code.
2) BcStTestDelegator: A single interface into product code non-public
members and types. This one is added to avoid multiple friendships
and to help to debug.
3) FakeReplicaBase: base class for destination and source fake replicas.
Defined for code reusability.
4) FakeDestination: a new class that fake ST replica. Used when we
would like to check a regular backup replica (source).
5) FakeSources: The new name of MockedSources, refactored to inherites
FakeReplicaBase.