test_app_state: add API peekBlock(), fix/modify getBlockAsync()
peekBlock() is used to get a block for viewing, without copying it.
We also fix getBlockAsync(): we are not interested in real
multithreading. The actual getBlock is done in a single thread, and a
future is artificially created and returned to the test body after the
result has been captured.