BCStateTran: modify computeNextBatchToFetch to match BlocksBatchDesc
rename computeNextRequiredBlock to computeNextBatchToFetch.

We need to calculate the next fetch batch. This will be done every time
ST starts getting missing blocks, or when a batch is over.
The main idea here is to have many fetch ranges in a batch. If a fetch
the range is larger than a batch, we just set the maximum batch size.

Here we also remove nextRequiredBlock_,nextBlockIdToCommit_.
The current code state won't compile.