RVBManager: add a detection mechanism -pruning is done during checkpoint
As of now, we choose not to assert for that, only warn.
This is not a fatal case. Worst case scenario: 50% of replicas start
pruning before the checkpoint and 50%  start pruning after.
This will be detected in another mechanism we are going to add: add
RVT state to checkpoint messages.